fix: suppression du dépôt imbriqué

+++ src/main/java/com/gestionStockBACKEND/gestionStockBACKEND/repository/UserRpository.java
@@ -4,5 +4,5 @@ package com.gestionStockBACKEND.gestionStockBACKEND.repository;
 import com.gestionStockBACKEND.gestionStockBACKEND.entity.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRpository extends JpaRepository<Users, Integer> {
+public interface UserRpository extends JpaRepository<Users, Long> {
 }
